Change native package name to runtime
