+++ graphql-native/src/main/java/io/ballerina/stdlib/graphql/runtime/schema/tree/Node.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.graphql.schema.tree;
+package io.ballerina.stdlib.graphql.runtime.schema.tree;
 
 import io.ballerina.runtime.api.types.Type;
