+++ graphql-native/src/main/java/io/ballerina/stdlib/graphql/runtime/utils/CallableUnitCallback.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.graphql.utils;
+package io.ballerina.stdlib.graphql.runtime.utils;
 
 import io.ballerina.runtime.api.async.Callback;
 import io.ballerina.runtime.api.values.BError;
