+++ graphql-native/src/main/java/io/ballerina/stdlib/graphql/runtime/utils/ModuleUtils.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.stdlib.graphql.utils;
+package io.ballerina.stdlib.graphql.runtime.utils;
 
 import io.ballerina.runtime.api.Environment;
 import io.ballerina.runtime.api.Module;
